rd-165 support callback as searcher
