+++ packages/robindoc/src/components/search/types.ts
@@ -1,0 +1,3 @@
+export type SearchItem = { title: string; href: string; description?: string };
+
+export type Searcher = (search: string, abortController: AbortController) => Promise<SearchItem[]>;
